pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_layer"
version = "0.1.0"
description = "Workspace for src and projects"
authors = [
    {name = "fddiferd", email = "fddiferd@gmail.com"}
]
requires-python = ">=3.10,<3.12"
dependencies = [
    "snowflake-snowpark-python[pandas]>=1.28.0,<2.0.0",
    "pandas>=2.2.3,<3.0.0",
    "scikit-learn>=1.5.0,<1.6.0",
    "matplotlib>=3.10.1,<4.0.0",
    "snowflake>=1.2.0,<2.0.0",
    "numpy>=1.23,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "cryptography>=41.0.0,<44.0.0",
    "snowflake-ml-python>=1.9.0,<2.0.0",
    "shap>=0.48.0,<0.49.0",
    "seaborn>=0.13.2,<0.14.0",
    "snowflake-telemetry-python>=0.7.1,<0.8.0",
    "joblib>=1.5.1,<2.0.0",
    "sqlparse>=0.5.3,<0.6.0",
    "ipykernel>=7.1.0",
    "torch>=2.9.1",
    "toml>=0.10.2,<0.11.0",
    "prophet>=1.1.1,<2.0.0",
]

[tool.setuptools.packages.find]
where = ["app"]
include = ["projects*", "src*"]

[tool.setuptools.package-dir]
"" = "app"
